refactor: unify card component
