+++ src/components/reviews/ReviewPanel.tsx
@@ -1,8 +1,11 @@
 import React, { type HTMLAttributes } from "react";
 import { cn } from "@/lib/utils";
-import Card from "@/components/ui/primitives/Card";
+import { Card } from "@/components/ui";
 
-export default function ReviewPanel({ className, ...props }: HTMLAttributes<HTMLDivElement>) {
+export default function ReviewPanel({
+  className,
+  ...props
+}: HTMLAttributes<HTMLDivElement>) {
   return (
     <Card
       aria-live="polite"
