--- src/components/ui/Card.tsx
@@ -1,49 +1,0 @@
-"use client";
-
-import * as React from "react";
-import PrimitiveCard from "./primitives/Card";
-import { cn } from "@/lib/utils";
-
-export const CardHeader = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn("flex flex-col space-y-1.5", className)} {...props} />
-  ),
-);
-CardHeader.displayName = "CardHeader";
-
-export const CardTitle = React.forwardRef<HTMLHeadingElement, React.HTMLAttributes<HTMLHeadingElement>>(
-  ({ className, ...props }, ref) => (
-    <h3 ref={ref} className={cn("text-lg font-semibold", className)} {...props} />
-  ),
-);
-CardTitle.displayName = "CardTitle";
-
-export const CardDescription = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<HTMLParagraphElement>>(
-  ({ className, ...props }, ref) => (
-    <p ref={ref} className={cn("text-sm text-muted-foreground", className)} {...props} />
-  ),
-);
-CardDescription.displayName = "CardDescription";
-
-export const CardContent = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn("pt-2", className)} {...props} />
-  ),
-);
-CardContent.displayName = "CardContent";
-
-export const CardFooter = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn("flex items-center pt-4", className)} {...props} />
-  ),
-);
-CardFooter.displayName = "CardFooter";
-
-const Card = React.forwardRef<HTMLDivElement, React.ComponentProps<typeof PrimitiveCard>>(
-  ({ className, ...props }, ref) => (
-    <PrimitiveCard ref={ref} className={cn("", className)} {...props} />
-  ),
-);
-Card.displayName = "Card";
-
-export default Card;
